Refactor to match established practices
Modify the README to match the format of other plugins

Modify API calls to match string constant formatting

Add error to notify when to create an issue
